+++ Tests/toolset.json
@@ -1,0 +1,10 @@
+{
+    "schemaVersion" : "1.0",
+    "linker" : {
+      "extraCLIOptions" : [
+        "--stack-first",
+        "-z", "stack-size=524288",
+        "--global-base=524288"
+      ]
+    }
+  }
